requirements.txt
Django==1.2.5
MySQL-python==1.2.3c1
South==0.7.3
django-activitysync==0.2.2
django-debug-toolbar==0.8.4
django-disqus==0.3.4
django-memcache-status==1.0.1
django-tagging==0.3.1
django-xmlrpc==0.1.2
feedparser==4.1
httplib2==0.6.0
oauth2==1.2.0
python-memcached==1.47
python-twitter==0.8.1
simplejson==2.1.2
wsgiref==0.1.2
yolk==0.4.1
docutils==0.6
Pygments==1.3.1
